Hand optimization

+++ cosmictiger/tree.hpp
@@ -226,14 +226,14 @@ struct cuda_kick_shmem {
 	union {
 		array<array<float, KICK_BLOCK_SIZE>, NDIM> f; // 384
 		struct {
-			array<int16_t, NITERS> count; // 8
-			array<array<int16_t, KICK_BLOCK_SIZE + 1>, NITERS> indices; //33
+			array<int, NITERS> count; // 8
+			array<array<int, KICK_BLOCK_SIZE + 1>, NITERS> indices; //33
 		};
 		array<float, KICK_BLOCK_SIZE> Lreduce;  // 4480
 	};
 	array<array<fixed32, KICK_PP_MAX>, NDIM> src;  // 3072
 	array<array<fixed32, MAX_BUCKET_SIZE>, NDIM> sink;  // 768
-	array<int8_t, MAX_BUCKET_SIZE> rungs; // 256
+	array<int, MAX_BUCKET_SIZE> rungs; // 256
 };
 
 struct kick_params_type {
